requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.61.1
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
